pyproject.toml
[tool.poetry]
name = "iwildcam"
version = "0.1.0"
description = ""
authors = ["LukeMathWalker <lucapalmieri1993@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
lazydata = "^1.0"
pillow = "^6.0"
click = "^7.0"
numpy = "^1.16"
pandas = "^0.24.2"
bidict = "^0.18.0"
matplotlib = "^3.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
explore = 'iwildcam.cli.explore:explore'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
